+ parallel execution for faster test runs
enabling support for parallel processing for faster runs @ saucelabs

– changed behave to unofficial version 1.2.4a1 from:
https://github.com/tokunbo/behave-parallel/tree/upstreamsync

this behave version was created on 10/30/13 and the pull request was
not accepted into the behave project for unknown reasons and the
functionality is still not in behave as of 11/20/14. This build has
been moderately tested and found to be working with no known issues.
However, due to the age and status of the pull request this is not an
ideal situation and a better long term solution is desired.

– updated setup.py to install the correct behave version when
installing from pip